Fix 'targetinfo' endpoint when iSCSI port is down
If iSCSI port is down or not listening the 'call_api'
will return an HTTP 503 code. In the context of
'targetinfo' endpoint this should not be considered
an error because if this happens we can assume that
there are no active sessions.

Signed-off-by: Ricardo Marques <rimarques@suse.com>
(cherry picked from commit 2dec0245678b323ff449a2943a53787018b8f83f)