Normalise gateway.conf controls on startup
There is a chance that an ses5.5 - 6 upgrade incorrectly stored int
values in the config as strings, and this can lead dashboard issues if
you want to make chances.

This patch will take a look at the config at startup, once it's loaded
from the rados pool, and if there is any ints stored as strings will
write correct them and write it back to the config on the pool.

While we're fixing ints in the config let's also go and
fix bsc#1163617

Fixes: bsc#1163617
Fixes: bsc#1164571
Signed-off-by: Matthew Oliver <moliver@suse.com>