Make settings mon config key store aware
When we are dealing with things like `trusted_ip_list` but ceph-iscsi is
running in a container modifying the list can be problematic. Sure the
config file is mounted into the container, but getting users to modify
that isn't very user friendly.

What would be better is if you could store the value in the mon config
key store. But then why stop at just the `trusted_ip_list`, ceph-iscsi's
settings are already fairly dynamic. So instead treat any value starting
with `config://` as a key for config key store.

This obviously wont work for the client name or the ceph conffile as we
need those to talk to the key store, but means anything else in the
config, if you so choose, can come from the dynamic ceph config store.

For example, if I had:

  trusted_ip_list = config://some/key/you/want/to/use

Then it'll look in the ceph config store under the key
`some/key/you/want/to/use`.

To set the key you just need to:

  ceph config key set some/key/you/want/to/use "192.168.0.1,192.168.0.2"

And then restart the container.

Signed-off-by: Matthew Oliver <moliver@suse.com>
(cherry picked from commit 3a12044b54e94b8bde1720b22c8bc7c29a779d3b)