Add a strip to ListSetting.Normalize
Currently when you specify a list in ceph-iscsi it _MUST_ be in the form
of 'item1,item2,item3'. Just delimitted by a comma (,). So if you
specify trusted_ip_list with spaces:

  trusted_ip_list: '10.1.1.1, 10.1.1.2'

This is actually fail as the space before 10.1.1.2 is retained.

This patch adds a strip() to the normalise call removing any extra
whitespace from the start and end of each item in the list.

Fixes: https://github.com/ceph/ceph-iscsi/issues/219
Signed-off-by: Matthew Oliver <moliver@suse.com>
(cherry picked from commit 2c792fc38d2fc0e940940f2feb3ca619920c1778)